Move decorator to component
