+++ packages/component/src/decorator/private/BorderLoader.module.css
@@ -1,21 +1,30 @@
 @keyframes borderAnimation-position {
   0% {
-    width: 33%;
     left: -33%;
+    width: 33%;
   }
 
   100% {
-    width: 33%;
     left: 100%;
+    width: 33%;
   }
 }
 
-:global(.webchat) .border-loader {
+:global(.webchat-fluent) .border-loader {
   border-radius: inherit;
-  position: relative;
 }
 
-:global(.webchat) .border-loader__loader {
+:global(.webchat-fluent) .border-loader__track {
+  --webchat-decorator-borderLoader-borderSize: var(--webchat-borderLoader-borderSize, 4px);
+
+  background-color: var(--webchat-colorNeutralBackground6);
+  height: var(--webchat-decorator-borderLoader-borderSize);
+  position: relative;
+  width: 100%;
+}
+
+
+:global(.webchat-fluent) .border-loader__loader {
   --webchat-decorator-borderLoader-color1: var(--webchat-borderLoader-color1,
       var(--webchat__border-animation--color-1));
   --webchat-decorator-borderLoader-color2: var(--webchat-borderLoader-color2,
@@ -24,7 +33,6 @@
       var(--webchat__border-animation--color-3));
   --webchat-decorator-borderLoader-color1Fade: var(--webchat-borderLoader-color1Fade, transparent);
   --webchat-decorator-borderLoader-color3Fade: var(--webchat-borderLoader-color1Fade, transparent);
-  --webchat-decorator-borderLoader-borderSize: var(--webchat-borderLoader-borderSize, 4px);
 
   animation: borderAnimation-position 3s linear infinite;
   background: linear-gradient(90deg,
@@ -34,7 +42,7 @@
       var(--webchat-decorator-borderLoader-color3) 92%,
       var(--webchat-decorator-borderLoader-color3Fade)) no-repeat;
   height: var(--webchat-decorator-borderLoader-borderSize);
-  inset: auto 0 0 0;
+  inset: 0;
   isolation: isolate;
   position: absolute;
   width: 100%;
