+++ packages/component/src/decorator/private/BorderLoader.tsx
@@ -1,17 +1,19 @@
-import React, { Fragment, memo, type ReactNode } from 'react';
-
+import React, { memo, type ReactNode } from 'react';
 import { useStyles } from 'botframework-webchat-styles/react';
+
 import styles from './BorderLoader.module.css';
 
-function Loader({ children }: Readonly<{ children?: ReactNode | undefined }>) {
+function BorderLoader({ children }: Readonly<{ children?: ReactNode | undefined }>) {
   const classNames = useStyles(styles);
 
   return (
     <div className={classNames['border-loader']}>
       {children}
-      <div className={classNames['border-loader__loader']} />
+      <div className={classNames['border-loader__track']}>
+        <div className={classNames['border-loader__loader']} />
+      </div>
     </div>
   );
 }
 
-export default memo(Loader);
+export default memo(BorderLoader);
