+++ packages/component/src/decorator/private/Decorator.tsx
@@ -1,16 +1,18 @@
 import { DecoratorComposer, type DecoratorMiddleware } from 'botframework-webchat-api/decorator';
 import React, { memo, type ReactNode } from 'react';
 
+import ThemeProvider from '../../providers/Theme/ThemeProvider';
 import BorderFlair from './BorderFlair';
 import BorderLoader from './BorderLoader';
 import createStyles from './createStyles';
-import ThemeProvider from '../../providers/Theme/ThemeProvider';
 
 const middleware: DecoratorMiddleware[] = [
   init =>
-    init === 'activity border' && (next => request => (request.state === 'completion' ? BorderFlair : next(request))),
+    init === 'activity border' &&
+    (next => request => (request.livestreamingState === 'completing' ? BorderFlair : next(request))),
   init =>
-    init === 'activity border' && (next => request => (request.state === 'informative' ? BorderLoader : next(request)))
+    init === 'activity border' &&
+    (next => request => (request.livestreamingState === 'preparing' ? BorderLoader : next(request)))
 ];
 
 const styles = createStyles();
@@ -20,7 +22,6 @@ function WebChatDecorator({ children }: Readonly<{ readonly children?: ReactNode
     <ThemeProvider styles={styles}>
       <DecoratorComposer middleware={middleware}>{children}</DecoratorComposer>
     </ThemeProvider>
-  );
-}
+  );}
 
 export default memo(WebChatDecorator);
