--- packages/fluent-theme/src/components/decorator/index.ts
@@ -1,1 +1,0 @@
-export { default as FluentThemeDecorator } from './private/Decorator';
