--- packages/fluent-theme/src/components/decorator/private/Decorator.tsx
@@ -1,20 +1,0 @@
-import { DecoratorComposer, type DecoratorMiddleware } from 'botframework-webchat-api/decorator';
-import React, { memo, type ReactNode } from 'react';
-
-import BorderFlair from './BorderFlair';
-import BorderLoader from './BorderLoader';
-
-const middleware: DecoratorMiddleware[] = [
-  init =>
-    init === 'activity border' &&
-    (next => request => (request.livestreamingState === 'completing' ? BorderFlair : next(request))),
-  init =>
-    init === 'activity border' &&
-    (next => request => (request.livestreamingState === 'preparing' ? BorderLoader : next(request)))
-];
-
-function FluentThemeDecorator({ children }: Readonly<{ readonly children?: ReactNode | undefined }>) {
-  return <DecoratorComposer middleware={middleware}>{children}</DecoratorComposer>;
-}
-
-export default memo(FluentThemeDecorator);
